+++ Thmanyah/Thmanyah/Domain/Entities/PodcastItem.swift
@@ -15,6 +15,6 @@ struct PodcastItem: Identifiable {
     let duration: Int?
     let language: String?
     let priority: Int?
-    let popularityScore: Double?
+    let popularityScore: Int?
     let score: Double?
 }
